chore(FormSheet): align TypeScript API with RNS
